+++ lib/libchen/test/src/base/utf8.hpp
@@ -15,13 +15,19 @@ TEST(BaseUtf8Test, General)
     if (conf::data.empty())
         return ::testing::internal::ColoredPrintf(::testing::internal::COLOR_YELLOW, "warning: you didn't specify test data folder, skip utf8 test\n\n");
 
+    // these txt files contains utf-8 chars, each char's size is 1byte ~ 6byte
     auto bom   = chen::fs::read(conf::data + "/utf8/bom.txt");
+    auto sub   = chen::fs::read(conf::data + "/utf8/sub.txt");
     auto code  = chen::fs::read(conf::data + "/utf8/code.txt");
     auto plain = chen::fs::read(conf::data + "/utf8/plain.txt");
 
-    EXPECT_EQ(13, chen::utf8::count(plain));
-    EXPECT_EQ(code, chen::utf8::substr(plain, 0, 1));
+    EXPECT_EQ(11, chen::utf8::count(plain));
+    EXPECT_EQ(sub, chen::utf8::substr(plain, 0, 7));
+    EXPECT_EQ("", chen::utf8::substr(plain, 5, 0));
     EXPECT_TRUE(chen::utf8::bom(bom));
     EXPECT_FALSE(chen::utf8::bom(plain));
+
+    EXPECT_EQ("a", chen::utf8::convert(0x0061));
+    EXPECT_EQ("©", chen::utf8::convert(0x00A9));
     EXPECT_EQ(code, chen::utf8::convert(0x9648));
 }